Split /auth into per-concern route/handler modules; clean up its paths
