+++ api/http/handlers/auth/index.ts
@@ -1,0 +1,17 @@
+export { getUser, login, logout, register } from './session';
+export { getGoogleOauthUrl, verifyGoogleOauth, googleIdTokenLogin } from './oauth';
+export { verifyEmail, resendEmailVerification } from './email-verification';
+export {
+  changePassword,
+  setPassword,
+  beginPasswordReset,
+  checkPasswordResetCode,
+  completePasswordReset,
+} from './password';
+export {
+  beginEmailChange,
+  getEmailChange,
+  getEmailChangeByCode,
+  cancelEmailChange,
+  completeEmailChange,
+} from './email-change';
